test: rename test suite lazy-loading to data-provider
