+++ packages/combo-box/test/data-provider-dynamic-size-polymer.test.js
@@ -1,4 +1,4 @@
 import '@vaadin/text-field/vaadin-text-field.js';
 import '../src/vaadin-combo-box.js';
 import '../src/vaadin-combo-box-light.js';
-import './dynamic-size.common.js';
+import './data-provider-dynamic-size.common.js';
