+++ packages/combo-box/test/data-provider-lit.test.js
@@ -1,6 +1,6 @@
 import './not-animated-styles.js';
-import './lazy-loading-styles.js';
+import './data-provider-styles.js';
 import '@vaadin/text-field/vaadin-lit-text-field.js';
 import '../theme/lumo/vaadin-lit-combo-box-light.js';
 import '../theme/lumo/vaadin-lit-combo-box.js';
-import './lazy-loading.common.js';
+import './data-provider.common.js';
